refactor: nettoyage routes spécifiques effectif
